refactor(client): page structure (#10)
