+++ packages/client/src/pages/NotePage.tsx
